+++ internal/ui/health_test.go
@@ -1,6 +1,7 @@
 package ui
 
 import (
+	"os"
 	"testing"
 	"time"
 
@@ -8,15 +9,20 @@ import (
 )
 
 func TestHealthService(t *testing.T) {
+
+	lockFile, err := os.CreateTemp("", "paw.lock")
+	require.NoError(t, err)
+	lockFile.Close()
+
 	t.Run("service not available", func(t *testing.T) {
-		status := HealthServiceCheck()
+		status := HealthServiceCheck(lockFile.Name())
 		require.False(t, status)
 	})
 
 	t.Run("service available", func(t *testing.T) {
-		go HealthService()
+		go HealthService(lockFile.Name())
 		time.Sleep(5 * time.Millisecond)
-		status := HealthServiceCheck()
+		status := HealthServiceCheck(lockFile.Name())
 		require.True(t, status)
 	})
 }